// src/modules/people/user/interfaces/user.interface.ts
export interface UserInterface {
  parent_id: null | number;
  name: string;
  mobile_number: string;
  password: string;
  pin: string;
  active: null | string;
  wrong_password: null | boolean;
  wrong_pin: null | boolean;
  risk_profiling: null | string;
  is_jompay_allowed: null | string;
  app_version: null | string;
  fcm_token: null | string;
  default_language: null | string;
  default_currency: null | string;
  default_country_id: null | number;
  applicant_id: null | string;
  check_id: null | string;
  application_id: null | string;
  document_id: null | string;
  is_on_fido: null | string;
  on_fido_update_count: null | number;
  on_fido_check_count: null | number;
  remember_token: null | string;
  father_name: null | string;
  mother_name: null | string;
  gender: null | string;
  marital_status: null | string;
  occupation: null | string;
  source_of_income: null | string;
  id_type: null | string;
  id_no: null | string;
  id_issue_country: null | string;
  passport_issue_date: null | string;
  id_expire_date: null | string;
  profile_photo: null | string;
  scan: null | string;
  scan_1: null | string;
  scan_2: null | string;
  date_of_birth: null | Date;
  permanent_address: null | string;
  city_id: null | number;
  province_id: null | number;
  country_id: null | number;
  post_code: null | string;
  present_address: null | string;
  present_city_id: null | number;
  present_province_id: null | number;
  present_country_id: null | number;
  present_post_code: null | string;
  note: null | string;
  nationality: null | string;
  status: null | string;
  blacklist: null | string;
  id_no_duplicate: null | string;
  created_by: null | number;
  managed_by: null | number;
  updated_by: null | number;
  deleted_by: null | number;
}